requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
whitenoise==6.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
python-decouple==3.8
Pillow==10.1.0
pandas==2.1.3
numpy==1.25.2
openpyxl==3.1.2
xlsxwriter==3.1.9
requests==2.31.0
azure-identity==1.15.0
msal==1.25.0
google-cloud-storage==2.10.0
google-cloud-logging==3.8.0
google-auth==2.23.4
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
scipy==1.11.4
python-dateutil==2.8.2
pytz==2023.3
django-environ==0.11.2
django-extensions==3.2.3
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
coverage==7.3.2
freezegun==1.2.2
responses==0.24.1
parameterized==0.9.0
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0
newrelic==9.2.0
django-health-check==3.17.0
django-silk==5.0.4
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-filter==23.4
django-import-export==3.3.1
django-mptt==0.15.0
django-taggit==4.0.0
python-magic==0.4.27
bleach==6.1.0
markdown==3.5.1
pygments==2.17.2
jsonschema==4.20.0
pyyaml==6.0.1
cryptography==41.0.7
jwt==1.3.1
pyjwt==2.8.0
oauthlib==3.2.2
social-auth-app-django==5.4.0
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
channels==4.0.0
channels-redis==4.1.0
asgiref==3.7.2
uvicorn==0.24.0
websockets==12.0
aioredis==2.0.1
django-storages==1.14.2
boto3==1.34.0
botocore==1.34.0
s3transfer==0.9.0
django-compressor==4.4
csscompressor==0.9.5
jsmin==3.0.1
htmlmin==0.1.12
django-cacheops==7.0.2
django-ratelimit==4.1.0
django-webpack-loader==2.0.1
django-widget-tweaks==1.5.0
sorl-thumbnail==12.10.0
easy-thumbnails==2.8.5
django-imagekit==5.0.0
pilkit==3.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
django-countries==7.5.1
babel==2.13.1
transliterate==1.10.2
unidecode==1.3.7
python-slugify==8.0.1
django-model-utils==4.3.1
django-timezone-field==6.1.0
django-jsonfield==3.1.0
jsonfield2==4.0.0.post0
django-picklefield==3.1.0
django-extra-fields==3.0.2
django-versatileimagefield==3.0
python-magic-bin==0.4.14
django-cleanup==8.0.0
django-bulk-update==2.2.0
django-bulk-sync==3.3.0
django-mass-edit==3.4.1
xlwt==1.3.0
tablib==3.5.0
defusedxml==0.7.1
et-xmlfile==1.1.0
lxml==4.9.3
beautifulsoup4==4.12.2
html5lib==1.1
webencodings==0.5.1
chardet==5.2.0
certifi==2023.11.17
urllib3==2.1.0
idna==3.6
charset-normalizer==3.3.2
six==1.16.0
packaging==23.2
setuptools==69.0.2
wheel==0.42.0
pip==23.3.1